requirements.txt
Pillow
colormap
easydev